pi4/pyproject.toml
[project]
name = "pideck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "pyside6>=6.9.1",
    "qt-material>=2.17",
]
